+++ src/client/vendor.js
@@ -4,6 +4,7 @@ import 'react';
 import 'react-dom';
 import 'react-router-dom';
 import 'react-intl';
+import 'react-tap-event-plugin';
 
 import 'history';
 import 'immutable';
@@ -11,10 +12,14 @@ import 'immutable';
 // Redux
 import 'redux';
 import 'react-redux';
+import 'connected-react-router';
 import 'connected-react-router/immutable';
 import 'redux-immutable';
 import 'redux-thunk';
 import 'redux-form/immutable';
 import 'react-intl-redux';
 
-import 'normalize.css';
+// material ui
+import 'material-ui/styles/MuiThemeProvider';
+import 'material-ui/RaisedButton';
+import 'material-ui/TextField';